Change null rendering in console to just `; null`
You can't MOLD a null, but several places in the code want to be able
to represent it somehow.  e.g. the console displays it (though it
suppresses the display of void).  This changes that display to just
the simple `; null`.

This doesn't work for places that want to render a null inline with
other values, e.g.:

    >> x: "not" y: null z: "good"

    >> -- x y z
    -- x: "not" y: ; null z: "good"

So this tries out the use of backslash in that case, as it is currently
an illegal character:

    >> -- x y z
    -- x: "not" y: \null\ z: "good"