Move %typespec.r descriptions into %types.r file
"Type specs" were an unfinished R3-Alpha concept, that when you said
SPEC-OF INTEGER! or similar, you would not just get a textual name for
it but optionally other information (like numeric limits).  Hence the
"spec" was an object whose fields were defined in %sysobj.r

The gist is reasonable, though in terms of being feature-focused, then
having arbitrary precision integers is more useful.

But the feature was never really developed, so all there was really was
the type's textual description.  This meant that %typespec.r was just
a really annoying file that would get out of sync and create problems.
(This commit noticed that a datatype called UNICODE! was mentioned,
which apparently has been lingering there unnoticed since R3-Alpha.
The recent killing of LIT-WORD! and LIT-PATH! also did not get noticed
so their descriptions were still there.)

So this merges typespec.r's description into types.r for easier upkeep.
The SPEC OF object is still generated from the description, for now,
as a placeholder to keep track of the feature--as much as it ever was.