Support for GET-BLOCK!, SET-BLOCK!, GET-GROUP!, SET-GROUP!
For description:

https://forum.rebol.info/t/1032

The new types are laid out in a pattern that makes it possible to
quickly transform between the GET/SET/"PLAIN" forms.  This is used to
make the change to the scanner a relatively minor tweak.

A GET-BLOCK! now acts like REDUCE.  Given the existence of VOID! and
the wide spectrum of places that produce voids in the system (e.g.
PRINT), this means REDUCE [PRINT "HI"] will give you a block with a
void in it, as will REDUCE [IF FALSE []].  Considering this can happen,
it seems that pre-emptively erroring on a reduce that produces null
is probably more annoying than just voidifying the null.  So this
goes ahead and does that, and lets it error later if that void gets
looked at or used...a policy matching Rebol2/R3-Alpha/Red with the
#[unset] concept (except not leading enumerations of the block to
believe they are holding an unset variable when they hold a void!)

This means eliminating the experimental /TRY and /OPT refinements,
which were a solution in search of a problem.  They suggest the need
perhaps for a REDUCE-EACH style routine, that lets a user easily build
their own form of REDUCE function a bit easier than calling EVALUATE
(though evaluate is really not hard to call)

Because SET-GROUP! and SET-BLOCK! do a relatively common thing with
their right-hand side, this goes ahead and factors out some code that
was already common to SET-PATH! and SET-GROUP! but was just being
repeated.