Simplify DUMP-OBJ (as SUMMARIZE-OBJ), fix description
The DESCRIPTION you would get in HELP for an ACTION! was not matching
the text you would get when searching by string, e.g. HELP "COLLECT"

https://github.com/metaeducation/ren-c/issues/919

This addresses that by using the same "digging" process to get at the
underlying description for derived action that plain HELP used.

In the process it also cleans up the unneccessarily obfuscated code
for DUMP-OBJ.  Changed name to SUMMARIZE-OBJ to avoid suggesting that
it is actually doing the output.