Simplify/shorten GET_VAL_FLAG()/etc., rename to CELL_FLAG
This turns GET_VAL_FLAG(), SET_VAL_FLAG(), CLEAR_VAL_FLAG(), and
NOT_VAL_FLAG() into macros that do token pasting, of the passed in
flag with VALUE_FLAG_##whatever.  This means you don't have to say
VALUE_FLAG_XXX at the callsite, and can just say XXX.

It means the flag must be literal and not stored in a variable, but
that is what 99.9% of the cases are anyway...and the macros are so
simple that you can just do what they do directly if the needs are
more complex.

Also, protections that were included to avoid the use of multiple flags
with the GET_VAL_FLAG() variant (and forcing you to use ALL_VAL_FLAGS()
or ANY_VAL_FLAGS() as appropriate) are no longer necessary.  There is
implicit protection, as the token paste can only be used for one flag.

While the ALL_VAL_FLAGS() and ANY_VAL_FLAGS() macros wouldn't work with
this method, there were basically zero usages of them.  So they are
removed.

Becase this touches nearly every callsite that uses a VALUE_FLAG_, it
goes ahead and does a name change to calling these "CELL_FLAG".  One
reason is that it makes the code look more literate, as CELL is an
entire word (one character shorter than VALUE).  But also, some of the
flags are used on cells that are not "values", e.g. they contain no
data and are merely formatted.  It's more technically accurate to call
these CELL_FLAGs.