Use shorthand strategy for EVAL_FLAG (formerly DO_FLAG)
This repeats the pattern applied to getting CELL_FLAG calls to be more
compact, by using token pasting.  In this case, frame flags had been
evolving for a long time without having any particular accessor,
because it seemed like it would just be longer:

    f->flag.bits & FRAME_FLAG_XXX
    GET_FRAME_FLAG(f, FRAME_FLAG_XXX)

But with the new notation it's shorter and cleaner:

    GET_EVAL_FLAG(f, XXX)

This gives an additional good argument for why it's better to call the
cell accessor GET_CELL_FLAG instead of GET_VAL_FLAG, because it frees
up the shorter and less ambiguous EVAL with less chance of mixing up.