Use shorthand token paste macro for GET_SERIES_FLAG()
The pattern of:

   GET_SER_FLAG(s, SERIES_FLAG_XXX);

Can read more literally and succinctly as:

   GET_SERIES_FLAG(s, XXX);

This does that transformation, using token pasting to hide the fact
that the flag definitions are actually not such short words as to
generate a collision with other things using that word.

It also makes the XXX a little more distinct, to aid in searching the
codebase for the terms (as jump-to-definition will no longer work due
to the use of this approach).

Variants are created for GET_ARRAY_FLAG() and GET_ACTION_FLAG(), to
make the callsites even briefer.

A side effect of this change is that the checks are stronger at both
compile time, and at run-time with certain debug settings, e.g.
GET_ACTION_FLAG() actually checks that what it has is a paramlist,
and GET_ARRAY_FLAG() checks that what it has is an array.