Adjust for more Spectre mitigation checks in MSVC
MSVC raises warnings when it would insert slowdowns on Spectre cases.
This addresses some of them by making changes that seem to make the
warnings go away.  That doesn't necessarily mean it makes the Spectre
vulnerabilities go away, but it means that we don't need to disable
the "treat warnings as errors" or ignore that warning.