APPEND-OF, MUTABLE, BIND bootstrap fixes
The Travis process is not correctly reporting certain bootstrap
failures.  This fixes three issues that were causing failure--stale
uses of APPEND-OF, a too-string MUTABLE, and a call of BIND with a
LIT-WORD! instead of a WORD! to look for a binding from.

(Because it's not clear what bind should accept, for the moment the
LIT-WORD! option is not used, vs. making it more tolerant.  That may
be subject to change.)