Eliminate BAR! datatype, add #postpone enfix setting
This replaces the idea of a separate BAR! datatype to implement | with
the idea that it is a "postponing-enfix invisible operator".  It thus
restores | to being a WORD! that anyone can override.

One consequence of this is that soft-quoting now can pick up | as an
ordinary value.  Since hard-quoting could already do it, that isn't
much of an issue.  The relative inconsequentialness of this concern
is outweighed by the advantages of not introducing a new datatype and
allowing more compatibility with historical Rebol, plus giving users
the ability to override it for alternative functionality.