New leftward-quoting rule: right hand wins unless at end
When left-quoting was introduced, it had the obvious problem of
arbitrating whether the left or the right would win in a "quote fight".

    >> plus-one: x => [print x + 1]  ; quoted word left to get x
    >> x-word: lit x  ; quoted right to get x

    >> lit => [print lit + 1]  ; ???

Both can't win.  The case above seems to be a good fit for letting =>
use lit as the name of its parameter.  But if the right always wins,
what about:

    >> help =>  ; people are vocal about being mad if there's no help

Because of Rebol's left to right nature, the initial approach started
out with trying to do a somewhat fuzzy priority scheme, where left
winning was the default...and then maybe soft quoting could be lower
priority than hard quoting.

That approach was complex, and it meant that understanding whether the
left or right would win was based on the specific pairing of
operations...vs the state of a known operation.  It also meant there
was complicated parameter checking, slowing the evaluator.

Having studied the cases, this goes with a much simpler rule: the
right always wins out for quoting the left *unless* the right is at
the end of an array.  This means that if you want to use something
like LIT to literally get a left quoting operator like =>, you have
too say e.g. `do [... (lit =>) ...]`.

The at-end exemption only applies when there is a WORD! or a PATH!
to the left of a left-quoting operator.  It would not make sense to
apply it in other cases, e.g. (1 left-quoting-op) has no reason to
defer to the 1, as it has no evaluator action and would just be
discarded.

Overriding this behavior can be done with (lit => |), as it really
is looking literally for an end.  So a barrier can pick up slack.