Add separate lookback cell to frame
The space-conscious nature of frames tried to be fairly minimal, and
only reserved one fixed cell of "workspace".  However, that cell was
doing double-duty as the place to store lookbacks from transient
feed sources, as well as being a general-purpose cell.  This
complicated things and made it not possible for the cell to be used
by dispatchers which may feed the frame forward--so they could no
longer take for granted one unit of temporary GC-protected space into
which they could evaluate.

Working around that concern became too annoying, and there are many
good reasons to not conflate the two purposes.  This goes ahead and
adds the extra cell, and reinstates its usage in things like CASE,
which simplifies that code.