Don't show tmp-boot.r in line numbering
There's actually some potential useful information in lines of code
attributed to tmp-boot.r, such as the line number of input entered
in the console.  However, it's not set up to be used well...and it
just makes error reports look cluttered.

Ideally the mezzanine code and boot code would have the correct file
names and be able to link you to GitHub to read the context.  But for
now anything tmp-boot.r is cleared out of the error before reporting.

Also goes to not using the unicode "(i)" symbol in the console.  It
is not available on Windows, and can otherwise make things look
weird.  Save this for user-configured skins (or some detection process
that the environment is good enough to show it).