Eliminate `bits` union from date year/month/day/zone
When R3-Alpha packs a datetime into a single cell, the time takes a
64-bit-pointer's worth of nanoseconds.  As the header is 32-bits on
32-bit platforms, that leaves 32 bits for the date portion.  So it
used bit fields.

In order to make comparisons "easier", the bit fields were aliased in
a union with a 32-bit integer.  This seemed to allow telling if one
date were greater than another with a single comparison.

That's not actually true about bit fields.  Although there was some
attempt to compensate for endianness, there's really little in the
C standard about how bit fields work.  Moreover, using a union to
alias access to a bit field structure as an integer is undefined
behavior.

This keeps the use of bit fields but just does hashing and comparison
manually.  It was in the way of another change, so this is isolated
out here...none of the existing tests failed...but more date/time
tests are needed.