Refactor header files, enable custom extension cell types
This is a major reorganization, that manages to cut the %sys-rebval.h
and %sys-value.h files down to a somewhat more manageable size.  It
also takes somewhat ancillary cell types (e.g. for MONEY! or GOB!
or STRUCT!) and makes it so that they can be defined "after-the-fact"
as cell payloads in a way that does not violate the C standard's
rules about strict aliasing.

It also uses token pasting to make access to the EXTRA and PAYLOAD
parts a bit easier to read, as well as to provide a hook point for
any future checks of these accesses in the future (such as to make sure
the same type is being read that was written prior).

In the process of this cleanup, several benefits arise:

* Undefined behavior of accessing bits for dates through a union pun
  is taken care of, by explicitly doing date field comparison.

* Endianness dependency for dates is removed

* Dependency on #pragma pack for events is removed

* Tuples can be one byte longer (to a full 8 bytes), and have a cleaner
  separation between the bytes and the length

The comments haven't all been vetted for being up-to-date and may have
some continuity errors due to pasting and rearrangment, but things
are more in the right place.