Enable COPY on DATE and TIME
Under rules of the TO functionality, a TO conversion of a type to
itself does the same thing as a COPY.  This is now enforced, and it
won't even try to run the COPY.

DATE and TIME did not have it enabled, so this adds that.  Other types
are being added on an as-needed basis to find out when anyone actually
copies those types, and what they expected to happen.