Switch REBGOB to be a REBARR type vs. its own pool
This commit is designed to help move GOB! out of the core and into an
extension--so that it will not take up space in builds that would not
be using it.  It also removes unwanted inter-dependencies between the
low level memory pools and GOBs.  In the process, it actually cleans
things up and resolves several issues and bugs endemic to problems
which the "optimized" code had.

...story is...:

The REBGOB datatype wants to hold references to Rebol values for both
GOB! "content" and "user data".  Gobs also hold references to other
gobs for parent/child relationships, and want these references to be
able to GC.  So there has to be some way of the garbage collector
knowing which pointers inside a GOB! are to Rebol REBSER* nodes (or
other gob nodes) in order to work.

In R3-Alpha, this was addressed by making the Gob structure something
similar to a REBSER node--it had a different size, and pointers in
different locations.  But it still used a pooled allocation, and the
pool could be marked and swept to do a GC.

But this introduced complexity, and that code didn't have the attention
put into it that the main REBSER GC had (for instance, avoiding crashes
on deep recursions).  Plus putting the GOB data into a custom structure
instead of fitting into something more like an OBJECT! meant that it
was harder to reuse code on it (for displaying, inserting into, etc.)

This switches REBGOB to being a REBARR, and the child panes list into
a BLOCK! of GOB!s.  This does make them bigger...however, it uses some
tricks to keep things in the realm of efficiency and still fit the
REBGOB into an 8-cell series pool (by being 7 elements long, it fits,
since it's 8 with the END).

The trick involves using some "custom" payload and extra bit patterns,
that have been generalized to be added after-the-fact and not break
any strict aliasing rules.  These allow using 3 of the 4 platform
pointers in that cell for arbitrary data...in GOB!s case that's two
floats for coordinates, and one extra for flags/misc, repeated over
four cells.  The savings are significant enough to be worth it, e.g.
if even one more cell had to be used then that one cell would double
the amount of memory used by all the GOB!s.  So packing is savvy.

In the non-packed cells, several bugs are resolved and issues are
simplified.  Previous attempts to collapse user-passed in series or
objects into a single REBSER* would lose things like index positions
of a string, or bindings of objects.

There are no current users of GOB! in Ren-C and a token few tests in
the test suite.  But to test this commit, old rebol-issues tickets
were looked up and added.  Some bugs were fixed by this commit, other
features were found to have been broken for a while and need work.
But this is a pretty solid proof of concept that the separation works.