If error during PATH! evaluation, show NEAR in PATH! format
Path evaluation changed to use ordinary REBFRM* as work on the model
went along.  One casualty of this was that for a few reasons, the
display of the "near" error message was showing a path rendered as
a BLOCK! when the error happened during path eval.

This corrects that issue--showing a partial amount of the path.  It
does not preserve whether it is a SET-PATH! or GET-PATH!, and it
perhaps should, and if it's too long say foo/bar/~~/(...): to indicate
there is more to come.