Get rid of rebRunInline() API
The concept behind rebRunInline was that you could execute a BLOCK!
without having to say `rebRun("do", block)`, which would add a stack
level for DO.  That would make debug traces show an extra function
call you didn't want.

But all it did was turn the thing you passed it into a GROUP!.  With
the existence of rebEval() to make elements "live" this eventually
turned into nothing special...and something you could do yourself
with the API.

This removes the unnecessary entry point, and updates the console code
that used it to make its own GROUP!s.