Postfix assertions, improve error reporting, invisible hijacks
This introduces the WAS, SO, and MATCHED postfix assertions--along
with the supporting function MATCHES which reverses the parameter order
of MATCH and returns a LOGIC!.

It also improves the HIJACK mechanic to work with invisibles, so that
assert (which is a hijacked function, with a verify implementation in
"debug mode") can be invisible.

Improves general error reporting, including having Apply_Only_Throws()
calls push their error up to a frame that may be more informative
to show the near information for.