Generate FIRST...TENTH functions in a loop, fix LAST
This uses a variadic looping construct to make the ordered pickers,
vs. having them all typed in by hand.  This makes it easier to adjust
all of them at once if any tweaks were made (wording in description,
for example).

It also improves LAST to be a specialization (so PICK reuses the
frame), and takes care of a quirk that was in the historical R3-Alpha
means of retriggering ACTION!s related to handling GOB!s.  Everything
can be handled now by the pick specialization