Add support for REFINEMENT! type hack to MATCH
The type system for ACTION! dispatch is not flexible enough at this
time to capture the idea of a datatype property.  e.g. you can't say
that a function only takes non-negative integers, you have to take
all INTEGER! and then filter inside the function.  There's no
NON-NEGATIVE-INTEGER! and no way to make it.

It's a desirable area for feature development, and for skippable
parameters discerning blank-headed paths  from ordinary PATH!s it is
essential.  Hence a hack for specifying type classes with an ISSUE!
was added, with one for #refinement!.  Then REFINEMENT!: #refinement!.

MATCH did not support this hack, so it was having bootstrap problems.
This adds the hack to MATCH.