Use goto in parse "optimization" to avoid indentation level
There's a strange claim of an optimization by second-guessing a range
of cases in a case statement that should be contiguous, and optimized
by the compiler...to short-cut doing the switch.  That shouldn't be
any faster with jump tables (and possibly slower).

This doesn't get rid of the optimization, but just changes it from an
if to a goto--which should be equivalent.  By using a goto it doesn't
add another indentation level, which was getting out of hand.

Also does some other minor cleanup and turning comments into assertions
in the outdented code.