COLLECT and KEEP in PARSE (with backtracking)
(Note: This is not particularly well tested, and is being committed
due to community demand...where it is expected that the community
will invest in making the tests for it!  Expect bugs, but the gist
seems to work.)

This implements COLLECT and KEEP keywords in PARSE.  They are justified
as features tailored to parse specifically because it is desirable
to have rule failures back out kept items.  For instance, with the
following code:

    parse [1 2 3] [
        collect x [
            keep integer! keep integer! keep text!
            |
            keep integer! keep [some integer!]
        ]
    ]

Without backtracking, this would give x as [1 2 1 2 3], when the likely
intended result was [1 2 3].

In addition to providing the backtracking (that Red did not with its
COLLECT), this changes it to take a variable to set, which is more
like SET and COPY... vs. affecting the overall return result of the
PARSE operation.  The PARSE still returns the series input as far
as it has advanced without failing, or null if it fails.

Following the pattern of existing "refinements" in R3-Alpha, the /ONLY
feature of KEEP is roughly implemented as KEEP ONLY vs. KEEP/ONLY:

    >> parse [1 2 3] [
         collect x [keep only [2 integer!] keep only integer!]
    ]
    == []

    >> x
    == [[1 2] [3]]

If it is desired to add material to what is being collected that does
not originate from the input series, a GET-BLOCK! can be used.  It
will be reduced and inserted (the reducing behavior helps keep it
more parallel to the expectations set in ordinary evaluation of a
GET-BLOCK!)

    >> parse [1 2] [
        collect x [keep integer! keep :['a <b>] keep integer!]]
    ]
    == []

    >> x
    == [1 a <b> 2]

Non-array series may be collected from also:

    >> parse "aaabbbccc" [
        collect x [keep [some "a"] some "b" keep [some "c"]]
    ]
    == ""

    >> x
    == ["aaa" "ccc"]

Nested collects are supported as well:

    >> parse [1 2 3 4] [
         collect a [
             keep integer!
             collect b [keep [2 integer!]]
             keep integer!
         ]
         end
    ]

    >> a
    == [1 4]

    >> b
    == [2 3]