Fix quoted FRAME! handling in GC
The GC was extracting the unescaped cell from the value it was passed
to mark.  However, it was holding this in a non-cell pointer.  This
meant that VAL_TYPE() checks were still allowed, which would return
a QUOTED! when the GC's intent was to mark whatever the underlying
quoting thing was.

This switches it to take an oddly named parameter (`quoted`) to
discourage its further use in the routine after the extraction, and
types the `v` local as a REBCEL*.