Kill cell EVAL_FLIP, eval EXPLICIT_EVALUATE, EVAL/ONLY
When the API was introduced, it was desired that spliced values act
as inert, so that you could write:

    const char *utf8 = rebSpell("mold", print_word);

Assuming that `print_word` contains the word PRINT, the idea here is
that it is not evaluative when spliced.  (If it were, then you'd get
an error about how PRINT was missing its argument, as it was trying to
be called vs. act as an inert WORD!.)

The first cut at doing this used VAL_FLAG_EXPLICIT_EVALUATE.  This mode
was originally made for supporting calls to Rebol routines from
internal APIs that (like the libRebol API) had "already evaluated"
arguments in its hand...and wanted to invoke functions without
evaluating them again.  (These functions predated frames, generic
quoting, and variadic feeds...so they didn't have options other than
a special flag the evaluator heeded.)

Yet the API wanted to evaluate *some* values (such as those loaded
from scanned portions).  So it needed a way to indicate the evaluation
be "flipped"...thus originating CELL_FLAG_EVAL_FLIP.  The idea was
that this was the same bit in the frame flags as EXPLICIT_EVALUATE,
so the evaluator could use that to its advantage in bit fiddling to
decide whether to evaluate or not faster.

This created a number of anomalies.  For instance: null values could
now be reified into frame feeds, because it was legal for these
"inert nulls" to be passed by the API.  This caused problems because
the arrays were not generally supposed to have nulls, so an exception
had to be made.

Similarly, there was concern about leaking this invisible "EVAL_FLIP"
flag into values that were resident in arrays.  It would mean a user
would have no visual indication of why a value was not running
evaluatively.  So there had to be special handling of this, as well
as special handling of those cases where the eval flip needed to be
preserved.

Generalized Quoting aimed to sweep away these concerns, by making the
evaluator *always* evaluative--and it is the job of whatever provides
the feed of values to add quote levels if it wishes to defeat that.
This reduces the amount of calculation that the evaluator has to do,
as it can just take the values it gets "as is".

This commit propagates the change, and does many necessary things to
keep the system working.  One big area that had to be reviewed was the
effects on APIs where previously cells that were nested in BLOCK!s
were not quoted, becoming quoted.  This necessitated pushing ahead
with the creation of the rebU() unevaluative function, and finalizing
the policy for quote escaping in APIs:

https://forum.rebol.info/t/1050

Other areas were just trying to coalesce two different code paths that
had been handling the variadic feed advancement into one.