Make C-DEBUG-BREAK an independently enableable
The C-DEBUG-BREAK native is very handy for causing a just-in-time
breakpoint in the point of evaluator source you want.  It was not
something that could be enabled independently from DEBUG_TICK_COUNT,
which meant it was not present in callgrind builds (callgrind does
not include any debug features that would affect performance, and
tick counting and tick breakpoint checking do).

This makes INCLUDE_C_DEBUG_BREAK native separate, and adds it to the
callgrind build configuration.