Kill native no-RETURN: trick, fixes requote/reskin
There was an optimization which does not actually put a RETURN cell in
the frame of natives, as they do not use definitional returns.  The
debug build had one to store the legal return types for the system to
check, but the release build assumed that natives always obey their
return type contracts.  (Bad return types from a native in the debug
build are an assert, not a conventional failure.)

This optimization meant that the release build did not have a RETURN:
field to read the `<requote>` status from.  So release builds would
say things like:

    >> find lit '''[a b c] 'b
    == [b c]  ; expected '''[b c]

Trying to work around this pointed out other problems of trying to
avoid having a cell, such as not being able to reskin something like
ADD to a version that didn't allow returning DECIMAL!

Weighing things in the balance, the optimization of not having the cell
is not worth it.  Natives can still avoid typechecking, but they will
have a RETURN cell in their frames that they won't use.  On the plus
side, this gives them a local for scratch space, which is GC safe...
and using that cell instead of an ordinary argument for temp work can
make usability in the debugger better (e.g. the original arg can be
left in place).