value/specifier/"gotten" into feed, instead of frame
Originally there was no separate abstraction for a "feed" from a
frame.  However, once there was, the only real reason to keep the
f->value, f->specifier, and f->gotten in the frame instead of the
feed was to keep from having to type f->feed->value/etc.

Keeping the f->value/f->gotten in sync was tricky to do and to
explain.  So push came to shove and this finally puts them into the
feed.  To help keep typing down, this adds a SHORTHAND macro which
uses a const pointer (e.g. the pointer doesn't change, vs. the data
pointed to) which aliases the fields.  This means one must use a *
to dereference them, but any decent optimizer will not actually
create the const pointer on the stack.

To avoid the potential pitfall of forgetting to indirect the pointer
(hence getting misleading logical boolean tests), the C++ build adds
a protection for disabling boolean conversions--where the dereference
operator simply provides the enclosed pointer reference.