Fix null parse rule molding error bug
The error raised on null parse rules was reporting the null value,
not the rule that produced the null.  The debug build was avoiding a
crash by printing out an ugly message (as molding nulls is somewhat
common in debug scenarios).  However, release builds would crash.

This adjusts the error call and adds a test.

(Includes fix for C++98 Travis build.)