Allow a parse recursion scoped to a single rule
R3-Alpha's parse recursions only worked on BLOCK!s.  The illusion of
them working with instances like `thru "a"` instead of `thru ["a"]`
came from a redundant function called "Parse_One_Rule".  It did a
subset of recognitions from the main PARSE, and would recurse on
a block.  But consequently, it did not cover all cases.

This changes the parse recursions to be able to operate on a common
instruction "feed".  That brings it closer to unification with the
evaluator (which is an ongoing goal, due to the desirability of having
common debug functionality for PARSE and regular evaluation...among
other reasons).  COLLECT and KEEP are both updated to use this, hence
you can now write things like:

     >> parse "aaabbb" [
        collect outer [
            some [collect inner keep some "a" | keep some "b"]
        ]
    ]
    == ""

    >> outer
    == ["bbb"]

    >> inner
    == ["aaa"]

COLLECT says "give me one unit of parse", so does KEEP. So it knows not
to go on and try and pick up the | as either part of `keep some "a"` or
`collect inner keep some "a"`.

Recursion in general has a cost associated, so some aspect of the
Parse_One_Rule fast matching is likely desirable to find a place for
(though hopefully not a redundant place).  Parse_One_Rule is tied into
the mechanics of TO and THRU--due to the fact that they need to
backtrack, and those are left as-is, for the moment...this is being
committed to fulfill a feature wish.