Add parallel JS for C Macros in Emscripten Build
Changes to the API to make fast C macros did not have a solution for
JavaScript.  Also, the feed separation commit broke the emscripten
build on Travis.

This resolves the breaks and patches issues so that the web REPL can
run again against libr3.js.  JavaScript stubs are added for the
macros...this is chosen vs. adding more entry points even though it
is less efficient, because it shows how the abilities would need to
be done by end users.  (Optimizing by adding entry points to the API
will be done later, once it's clear that the non-optimized ways all
are coherent.)