Move IMAGE! data type to extension
With IMAGE! changed to basically be a container which adds a size to
a backing store BINARY!, that alone takes care of much of what images
needed to do with encoding and decoding.

Functionality beyond that was semantically rather poorly defined:

https://github.com/rebol/rebol-issues/issues/801

Basically every line of code involving a scalar series position and
an image raised a semantic issue.  This takes all of that questionable
code and cuts it out of the core, so that all builds not be burdened
with those issues.

A consequences of moving image out of the core is that it becomes a
victim of the lack of definition for user-defined types.  This means
FOR-EACH does not have a way to work on IMAGE! directly--as it is not
a GENERIC.  However, the image data can be enumerated via the
BYTES OF accessor of its internal data.  Improvements to this should be
pursued by improving user-defined types in general.