Return COLLECT to empty block default, chains on COLLECT*
It is technically desirable for COLLECT to be built on a null-returning
foundation.  It has better performance and memory use characteristics
to not allocate a series node and array memory block if it is not
needed.  Additionally, many clients can benefit from discerning a
"nothing was collected" case, and being null means the result is both
logically false and able to trigger reactions from THEN/ELSE/ALSO.

However, the block form is also useful.  It is also arguable that the
distinct result of "nothing collected" may be usefully discerned from
having a blank input to the collect, to chain out a NULL.

For these reasons, this gives the empty-block returning COLLECT the
primary name--while still building it on the null-returning form.
The null returning form is tentatively named COLLECT*, which parallels
TAKE* as the "low level take" that returns null vs. erroring when
nothing is available to take.