Prep IMAGE! for removal from core to extension
This removes accessors like VAL_IMAGE_HEAD() from the core, so that
extensions which want to work with images do so through the libRebol
API, with `make image!` (like any other client would).  To enable this,
the IMAGE! will give a BINARY! of its rgba data back with `bytes of`.

That required changing the IMAGE! to use an ordinary BINARY! for its
payload, which means that things like the PNG codec can now directly
use rebRepossess() on the memory buffer to get a binary, and not need
to copy the decoded image into a series with length 4 elements.

Because it is switching to a "Custom" payload, this also removes IMAGE!
from the ANY-SERIES! class.  (There are many questions about its
membership in that class in the first place.)