Fix DEFLATE to return DEFLATE data, not GZIP
The wrapper for zlib translates Rebol symbols like ZLIB or GZIP to
what envelope format to use.  The envelope format request for getting
no envelope (e.g. raw deflate) uses SYM_NONE, and not nullptr.  This
is for technical reasons--because the bootstrap process itself has
to do decompression prior to any symbols being decompressed and
loaded, and GZIP is used for that.

libRebol APIs took this into account, but the DEFLATE and INFLATE
natives were passing in nullptr when they meant to pass in SYM_NONE.
This went unnoticed because DEFLATE and INFLATE was used in paired
situations, that is only revealed if used with an external source.
However, trying to unzip files with command-line utilities that were
built with unzip.reb exposed the flaw.